Fixed issue where tags at the beginning of summary not parsed correctly.
If tags appeared before the category when parsing, the tags would not be
added to the list of tags in the commit message. This was due, in part, to
a regular expression that did not consider the possibility that the tag
could be at the beginning of a line, and partially due to the fact that
tags were only parsed from the remainder of the line AFTER categories had
been removed.

Fixes #928.